[NETBEANS-895] Fixing adding imports when completing field initializer.
