+++ java.editor/test/unit/data/goldenfiles/org/netbeans/modules/java/editor/completion/JavaCompletionProviderTest/1.8/testAnnotation.pass2
@@ -1,0 +1,10 @@
+package test;
+
+import java.util.Set;
+
+@Annotation(Set.class)
+public @interface Annotation {
+
+    public Class<java.util.Set> value();
+
+}
