+++ java.editor/test/unit/data/goldenfiles/org/netbeans/modules/java/editor/completion/JavaCompletionProviderTest/1.8/testFieldInitializer.pass2
@@ -1,0 +1,10 @@
+package test;
+
+import java.util.HashSet;
+import java.util.Set;
+
+public class FieldSet {
+
+    public Set<String> field = new HashSet<String>;
+
+}
