+++ java.editor/test/unit/data/org/netbeans/modules/java/editor/completion/data/Annotation.java
@@ -1,0 +1,8 @@
+package test;
+
+@Annotation()
+public @interface Annotation {
+
+    public Class<java.util.Set> value();
+
+}
