+++ java.editor/test/unit/data/org/netbeans/modules/java/editor/completion/data/FieldSet.java
@@ -1,0 +1,9 @@
+package test;
+
+import java.util.Set;
+
+public class FieldSet {
+
+    public Set<String> field;
+
+}
