+++ java.editor/test/unit/src/org/netbeans/modules/java/editor/completion/JavaCompletionItemPerformTest.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.netbeans.modules.java.editor.completion;
+
+/**
+ *
+ * @author Jan Lahoda, Dusan Balek
+ */
+public class JavaCompletionItemPerformTest extends CompletionTestBase {
+
+    public JavaCompletionItemPerformTest(String testName) {
+        super(testName);
+    }
+
+    public void testFieldInitializer() throws Exception {
+        performTest("FieldSet", 91, " = new HashS", "HashSet", "testFieldInitializer.pass2");
+    }
+
+    public void testAnnotation() throws Exception {
+        performTest("Annotation", 27, "Se", "Set", "testAnnotation.pass2");
+    }
+}
